Use single page size helper with new larger values
Reduce duplicated code with single helper for pagination

On larger deployments we should allow for larger pages
